// src/app/interfaces/candidate.ts
export interface Candidate {
    address_line: string;
    asking_benefits: string;
    asking_salary: number;
    asking_salary_type: number;
    bookmarked: number;
    can_rate_stars: number;
    can_view_profile: number;
    chat_group_id: number;
    city_name: string;
    created_at: Date;
    description: string;
    email: string;
    first_name: string;
    id: number;
    last_name: string;
    profile_picture:string;
    state_name: string;
    status: number;
    total: number;
    total_percent: number;
    total_point: number;
    enable_show_avatar: number;
}